+++ frontend/correctink/lib/app/screens/edit/modify_profile.dart
@@ -6,6 +6,7 @@ import 'package:correctink/widgets/widgets.dart';
 import 'package:go_router/go_router.dart';
 import 'package:localization/localization.dart';
 
+import '../../../widgets/buttons.dart';
 import '../../data/models/schemas.dart';
