+++ frontend/correctink/lib/app/screens/inbox_page.dart
@@ -9,6 +9,7 @@ import 'package:provider/provider.dart';
 
 import '../../blocs/messages/message_list.dart';
 import '../../utils/router_helper.dart';
+import '../../widgets/buttons.dart';
 import '../../widgets/widgets.dart';
 import '../data/models/schemas.dart';
 import '../data/repositories/collections/users_collection.dart';
