+++ frontend/correctink/lib/app/screens/learn_page.dart
@@ -10,6 +10,7 @@ import 'package:provider/provider.dart';
 import '../../utils/learn_utils.dart';
 import '../../utils/router_helper.dart';
 import '../../utils/utils.dart';
+import '../../widgets/buttons.dart';
 import '../../widgets/snackbars_widgets.dart';
 import '../../widgets/widgets.dart';
 import '../data/models/schemas.dart';
