+++ frontend/correctink/lib/app/screens/log_in.dart
@@ -5,6 +5,7 @@ import 'package:localization/localization.dart';
 import 'package:provider/provider.dart';
 
 import '../../utils/router_helper.dart';
+import '../../widgets/buttons.dart';
 import '../../widgets/widgets.dart';
 import '../data/app_services.dart';
