+++ frontend/correctink/lib/blocs/app_bar.dart
@@ -5,6 +5,7 @@ import 'package:go_router/go_router.dart';
 import 'package:provider/provider.dart';
 
 import '../utils/router_helper.dart';
+import '../widgets/buttons.dart';
 
 class CorrectInkAppBar extends StatefulWidget implements PreferredSizeWidget {
   const CorrectInkAppBar({required this.backBtn, Key? key}) : super(key: key);
