+++ frontend/correctink/lib/blocs/sets/card_exist_dialog.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:localization/localization.dart';
 
+import '../../widgets/buttons.dart';
 import '../../widgets/widgets.dart';
 
 enum CardExistChoice {
