+++ frontend/correctink/lib/blocs/sets/set_sorting.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:localization/localization.dart';
 
+import '../../widgets/buttons.dart';
 import '../../widgets/widgets.dart';
 
 enum SetSortingField {
