+++ frontend/correctink/lib/blocs/tasks/reminder_widget.dart
@@ -9,6 +9,7 @@ import 'package:intl/intl.dart';
 import 'package:localization/localization.dart';
 
 import '../../utils/utils.dart';
+import '../../widgets/buttons.dart';
 
 
 enum RepeatMode{
