+++ frontend/correctink/pubspec.yaml
@@ -3,7 +3,7 @@ description: CorrectInk is a study app with tasks and flashcards.
 
 publish_to: "none" # Remove this line if you wish to publish to pub.dev
 
-version: 1.4.0
+version: 1.4.1
 
 environment:
   sdk: ">=3.0.0"
@@ -29,7 +29,6 @@ dependencies:
   localization: ^2.1.0
   rxdart: ^0.27.7
   timezone: ^0.9.2
-  flutter_timezone: ^1.0.7
   permission_handler: ^10.4.3
   markdown: ^7.1.1
   flutter_markdown: ^0.6.17
